Fix: Add RecaptchaEnterpriseCallbackFailureTest for comprehensive failure handling scenarios
